src/main/java/com/tecnico/teste/cadastro/usuarios/user/controller/UserController.java
package com.tecnico.teste.cadastro.usuarios.user.controller;

import com.tecnico.teste.cadastro.usuarios.user.dto.UserDTO;
import com.tecnico.teste.cadastro.usuarios.user.dto.UserUpdateDTO;
import com.tecnico.teste.cadastro.usuarios.user.service.UserServiceManager;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("users")
@RequiredArgsConstructor
public class UserController {

    private final UserServiceManager userService;

    @GetMapping
    public ResponseEntity<List<UserDTO>> findAll() {
        return ResponseEntity.ok(userService.findAll());
    }

    @GetMapping("/{id}")
    public ResponseEntity<UserDTO> findById(@PathVariable("id") Long id) {
        return ResponseEntity.ok(userService.findById(id));
    }

    @PostMapping
    public ResponseEntity<UserDTO> save(@Valid @RequestBody UserDTO dto) {
        return new ResponseEntity<>(userService.save(dto), HttpStatus.CREATED);
    }

    @PutMapping("/{id}")
    public ResponseEntity<UserDTO> update(@PathVariable("id") Long id, @Valid @RequestBody UserDTO dto) {
        return ResponseEntity.ok(userService.update(dto, id));
    }

    @PatchMapping("/{id}")
    public ResponseEntity<UserDTO> merge(@PathVariable("id") Long id, @Valid @RequestBody UserUpdateDTO dto) {
        return ResponseEntity.ok(userService.merge(dto, id));
    }

    @DeleteMapping
    public ResponseEntity<HttpStatus> delete(@RequestBody UserDTO dto) {
        userService.delete(dto);
        return ResponseEntity.noContent().build();
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<HttpStatus> deleteById(@PathVariable("id") Long id) {
        userService.deleteById(id);
        return ResponseEntity.noContent().build();
    }

}
